Services/api/src/config.py
import os
from pathlib import Path

from dotenv import load_dotenv

# Local .env file
PATH_ENV = Path(__file__).parent.parent / ".env"

load_dotenv(PATH_ENV)

REDIS_HOST = os.environ.get("REDIS_HOST")
REDIS_PORT = os.environ.get("REDIS_PORT")

POSTGRES_DB = os.environ.get("POSTGRES_DB")
POSTGRES_USER = os.environ.get("POSTGRES_USER")
POSTGRES_PASSWORD = os.environ.get("POSTGRES_PASSWORD")
DB_HOST = os.environ.get("DB_HOST")
DB_PORT = os.environ.get("DB_PORT")
