Services/api/pyproject.toml
[tool.poetry]
name = "service-api"
version = "0.1.0"
description = ""
authors = ["FlameInBrainSerj <Krivosheev.sergey.01@mail.ru>"]
readme = "README.md"

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cpu"
priority = "supplemental"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.0"
torch = "^2.2.1"
torchtext = "^0.17.1"
pandas = "^2.2.1"
numpy = "^1.26.4"
celery = "^5.3.6"
flower = "^2.0.1"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.28"
asyncpg = "^0.29.0"
starlette-exporter = "^0.21.0"
selenium = "^4.18.1"
natasha = "^1.6.0"
nltk = "^3.8.1"
h5 = "^0.9.2"
python-multipart = "^0.0.9"
types-redis = "^4.6.0.20240311"
gunicorn = "^21.2.0"
fastapi-cache2 = {extras = ["redis"], version = "^0.2.1"}
uvicorn = "^0.28.0"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
black = "^24.2.0"
flake8 = "^7.0.0"
mypy = "^1.9.0"
pre-commit = "^3.6.2"
flake8-bugbear = "^24.1.17"
flake8-builtins = "^2.2.0"
flake8-commas = "^2.1.0"
pep8-naming = "^0.13.3"


[tool.poetry.group.test.dependencies]
pytest = "^8.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
